Revert "chore: add comparator for diff tester (#5002)"
This reverts commit 8afe46b16850a8833b9812c7955d6a7653cd2ef7.
