+++ .github/workflows/release-nightly.yml
@@ -61,7 +61,7 @@ jobs:
           path: artifacts
 
       - name: Build node packages
-        run: pnpm run build:js && pnpm run build:viewer
+        run: pnpm run build:js
 
       - name: Move artifacts
         run: node scripts/build-npm.cjs
