+++ .github/workflows/release.yml
@@ -79,7 +79,7 @@ jobs:
           path: artifacts
 
       - name: Build node packages
-        run: pnpm run build:js && pnpm run build:viewer
+        run: pnpm run build:js
 
       - name: Move artifacts
         run: node scripts/build-npm.cjs
