+++ packages/rspack-test-tools/src/compare/index.ts
@@ -2,4 +2,3 @@ export * from "./replace-module-argument";
 export * from "./replace-runtime-module-name";
 export * from "./format-code";
 export * from "./compare";
-export * from "./comparator";
