+++ packages/rspack-test-tools/src/index.ts
@@ -2,5 +2,3 @@ export * from "./type";
 export * from "./test/tester";
 export * from "./processor";
 export * from "./reporter";
-export * from "./compare";
-export * from "./plugin";
