--- packages/rspack-test-tools/src/plugin/index.ts
@@ -1,2 +1,0 @@
-export * from "./rspack-diff-config-plugin";
-export * from "./webpack-diff-config-plugin";
